Probability Modifiers
Change the logic to how Lineups are generated to include a notion of PositionProbability. These are attributes of a Player, and are associated with the player's Skills. These probabilities are based on the skills, but have randomness to them to create chance in assignments.

Update the core logic flow to use this new concept, removing the old code. Include a readme to help describe the algorithm.